Refactor: DB <-> Application 간 데이터 변환하는 컨버터 위치 수정
