+++ src/main/java/com/example/parking/infra/converter/FeeTypeConverter.java
@@ -1,5 +1,6 @@
-package com.example.parking.domain.searchcondition;
+package com.example.parking.infra.converter;
 
+import com.example.parking.domain.searchcondition.FeeType;
 import jakarta.persistence.Converter;
 
 @Converter(autoApply = true)
