+++ src/main/java/com/example/parking/infra/converter/OperationTypeConverter.java
@@ -1,4 +1,4 @@
-package com.example.parking.domain.searchcondition;
+package com.example.parking.infra.converter;
 
 import com.example.parking.domain.parking.OperationType;
 import jakarta.persistence.Converter;
